+++ src/main/java/ru/sfedu/hiber/lab4/set/model/Category4Set.java
@@ -18,11 +18,11 @@ public class Category4Set implements Serializable {
     private String name;
 
     @Column(name="prod_id")
-    private int productId;
+    private long productId;
 
     public Category4Set() { }
 
-    public Category4Set(String name, int productId) {
+    public Category4Set(String name, long productId) {
         this.name = name;
         this.productId = productId;
     }
@@ -43,11 +43,11 @@ public class Category4Set implements Serializable {
         this.name = name;
     }
 
-    public int getProductId() {
+    public long getProductId() {
         return productId;
     }
 
-    public void setProductId(int productId) {
+    public void setProductId(long productId) {
         this.productId = productId;
     }
